Ramdisk: Bring back Flash profile (version 1.1)
This profile did really well with the masses apparently and the lag
that I was experiencing that I thought made it bad was really a ROM
related issue... so bring it back! I have tuned the values to be
slightly more aggressive across the board since I want the profile to
be bursty.

Design goals
============

The entire goal of this profile is bringing power back a little so
we save battery but not neutering performance in the process. We
want to harness what we are given.

Since we have two different set of cores, we need to treat them a
bit differently based out their strengths and weaknesses. The little
set of cores (A53) use a lot less power than the big set (A57); the
maximum frequency of the little cores (1555MHz) uses less power than
the minimum (stock) frequency of the big cores (633MHz). As a result,
if we can avoid it, we really want to avoid using the big cluster.

However, if the big cluster needs to be worked, we cannot starve it
if we want to avoid lag (which we do). Little was not designed to do
everything, that defeats the purpose of big.LITTLE in the first place.

Little's profile
================

I allow the CPU to flow easily between 633-1248MHz with a somewhat
hard stop at 1344MHz (same as stock) then allow more scaling to 1478
and 1555 (even to overclocked values) as necessary. Input boost is
enabled at 1248MHz for 60 milliseconds. I don't ever want to see this
profile lag so keeping the initial input boost high will prevent it
from doing so. It'd be nice if there was a 1000-1100MHz value because
this does feel a little aggressive but the governor will allow the
frequency to drop down if the target load is low enough after the
initial boost. Little is designed to be used aggressively so we want
to give it all the work we can. Google's stock values are input boost
at 1344MHz for 40ms with target loads being 90 almost across the
board. This profile should compete with that performance while giving
better battery life.

Big's profile
=============

I want to cap the processor at 633MHz with a high initial target
load of 95, as this will prevent the processor from jumping to a
value that it doesn't need to. Only time the load should jump higher
than this would be during app launches and gaming. Things like typing
or scrolling should all be handled by little. However, if the load is
greater than 95, I dropped the loads up until 1248MHz, which should
allow it to flow easily up to that frequency. After that, it is still
sort of hard to climb up but is possible if the target load is that
hard. Other thing to note is we are taking only one target load
sample at a time, whereas with little we are free to take two at a
time. This will make big more responsive because we want it to drop
quickly. With little's power consumption being so low, we can feel
free to let it hang a little bit to make sure everything is being
done properly.

TL;DR
=====

This profile is about abusing little while letting big do its
job when necessary, which will result in a quick phone while
still giving decent battery savings over the stock profile.

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>