Include prebuilt Substratum
This is a debug APK I have compiled myself, meaning the Play Store
version will be completely incompatible with this due to mismatching
keys. If you are a ROM with users other than yourself, please replace
this APK with one from APK Mirror that is signed with the Play Store
keys so users can still take updates.

Link: http://www.apkmirror.com/apk/projekt/substratum-theme-engine/

Additionally, because I build for this fuccboi (@nicholaschum), add
a way to disable including the APK via a bash export:

export HAS_SUBSTRATUM=false

Signed-off-by: Nathan Chancellor <natechancellor@gmail.com>